// models/verifiable-credentials/types/VerifiableCredential.d.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

/**
 * A [[VerifiableCredential]] is fact-checking review/attestation of claims made about a credentialSubject
 */
export interface VerifiableCredential {
  "@context": string[];
  type: string[];
  id: string;
  issuer: {
    id: string;
    ethereumAddress?: string;
    [k: string]: unknown;
  };
  credentialSubject: {
    id: string;
    ethereumAddress?: string;
    [k: string]: unknown;
  };
  credentialSchema: {
    id: string;
    type: string;
  };
  issuanceDate: string;
  expirationDate?: string;
  proof?: {
    type: string;
    verificationMethod?: string;
    ethereumAddress?: string;
    created?: string;
    proofPurpose?: string;
    proofValue?: string;
    eip712?: {
      domain: {
        chainId: number;
        name: string;
        version: string;
        verifyingContract?: string;
      };
      types: {
        [k: string]: unknown;
      };
      primaryType: string;
    };
    [k: string]: unknown;
  };
  evidence?: {
    id: string;
    type?: string[];
    verifier?: string;
    evidenceDocument?: string;
    subjectPresence?: string;
    documentPresence?: string;
  };
  credentialStatus?: {
    id: string;
    type: string;
  };
}